# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "What to watch API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user data",
                "description": "Get user data",
                "operationId": "a65897868f1d4105404848067624d207",
                "responses": {
                    "200": {
                        "description": "List of films",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get user data",
                "description": "Get user data",
                "operationId": "07961faf38790d89b63ebf00b8cabb58",
                "parameters": [
                    {
                        "name": "_method",
                        "in": "query",
                        "example": "PATCH"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of films",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/user/avatar": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Save user avatar",
                "description": "Save user avatar",
                "operationId": "a4c7ce6e96fe7d0c733a2cdb99b8df4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveOrReplaceUserAvatarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film with specified id not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "avatarPath": {
                                            "type": "string",
                                            "example": "ef7d1976-5cd5-4c99-9bdf-cbd2209f214e/avatar/KgoFL8KpEtajLXJ225JjcMBIbMlKbXVd.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PatchUserRequest": {
                "properties": {
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "Ivan",
                        "nullable": false
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "ivan@somemail.ru",
                        "nullable": false
                    },
                    "password": {
                        "description": "Password",
                        "type": "string",
                        "example": "ud5R#gh78Qi!6",
                        "nullable": false
                    },
                    "password_confirmation": {
                        "description": "Password confirmation",
                        "type": "string",
                        "example": "ud5R#gh78Qi!6",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "SaveOrReplaceUserAvatarRequest": {
                "properties": {
                    "file": {
                        "description": "User avatar (png, jpg, jpeg, svg)",
                        "type": "file",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Ivan"
                    },
                    "email": {
                        "type": "string",
                        "example": "ivan@somemail.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "User"
                    },
                    "avatarPath": {
                        "type": "string",
                        "example": "ef7d1976-5cd5-4c99-9bdf-cbd2209f214e/avatar/KgoFL8KpEtajLXJ225JjcMBIbMlKbXVd.jpg"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}